refactor(base_messages): split player_init/base_messages along variant-handler seams
Tech-debt #529 file-size split (no behavior change).

base_messages/mod.rs (740) → thin BaseToCellMsg match (355) delegating each
variant family to a sibling submodule:
- lifecycle.rs    — CreateEntity / DestroyEntity / ConnectEntity / DisconnectEntity
- movement.rs     — EntityMove (+ movement_reject_label + sampling counter)
- inventory_events.rs — InventoryItemMoveApplied / Removed / Granted / ItemUsed
- ability_granted.rs  — AbilityGranted (hotbar refresh + trainer resend)
- minigame.rs     — MinigameResult
- gm_spawn.rs     — GmSpawnNpcReady

player_init.rs (1387) → player_init/ directory:
- mod.rs (493)    — handle_init_player_state + send_known_abilities_update
- tests.rs (900)  — the three inline #[cfg(test)] mod test blocks

Behavior-preserving · re-export discipline · zero public-surface change
(handle_base_message stays pub(super); message_loop.rs call site unchanged).

Part of #529
